fix bug handle order item, choose table

+++ src/main/java/com/something/restaurantpos/repository/IMenuCategoryRepository.java
@@ -3,5 +3,5 @@ package com.something.restaurantpos.repository;
 import com.something.restaurantpos.entity.MenuCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IMenuCategoryRepository extends JpaRepository<MenuCategory, Long> {
+public interface IMenuCategoryRepository extends JpaRepository<MenuCategory, Integer> {
 }
